% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustergene.R
\name{clustergene}
\alias{clustergene}
\title{Cluster genes}
\usage{
clustergene(genefit, clunum = NULL, maxclu = 20)
}
\arguments{
\item{genefit}{Direct output from calling \code{genefit()}.}

\item{clunum}{A positive integer indicating number of clusters. If \code{NULL}, will be automatically selected by \code{findPC}.}

\item{maxclu}{A positive integer indicating the maximum number of clusters to be considered by \code{findPC}. Not used if \code{clunum} is not \code{NULL}.}
}
\value{
A list of gene clusters and hclust.
}
\description{
Assign clusters for genes.
}
\examples{
pt <- setNames(1:100, paste0('cell',1:100))
cd <- matrix(c(rep(1:10,10),rep(c(1:5,11:15),each=10)),ncol=2,dimnames=list(paste0('cell',1:100),c('x','y')))
res <- paella(pt,cd)
mat <- matrix(rnorm(1000),nrow=10,dimnames=list(paste0('gene',1:10),paste0('cell',1:100)))
fit <- genefit(mat,res)
clu <- clustergene(fit,maxclu=10)
}
\author{
Wenpin Hou<wh2526@cumc.columbia.edu>, Zhicheng Ji
}
