% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftest.R
\name{difftest}
\alias{difftest}
\title{Testing differential gene expression dynamics}
\usage{
difftest(mat, paella, genecut = 0.05, ncores = 10)
}
\arguments{
\item{mat}{A gene expression matrix where each row is a gene and each column is a cell/spot. Row names are gene names and column names are cell/spot names. The column names must be present when fitting \code{paella()}.}

\item{paella}{Direct output from calling \code{paella()}.}

\item{genecut}{A numeric value between 0 and 1 indicating the gene filtering cutoff. Genes that have expression larger than median of all gene expression values in at least \code{genecut} cells are preserved.}

\item{ncores}{A positive integer of the number of cores to be used for parallel computing. If 1, no parallel computing will be done.}
}
\value{
A data frame of FDR and pval of overall, mean, and trend differential tests, as well as the type of differential for each gene.
}
\description{
Testing if the temporal gene expression dynamics along pseudotime is differential across spatial sub-trajectories for each gene. The tests involve overall, mean, and trend differential.
}
\examples{
pt <- setNames(1:100, paste0('cell',1:100))
cd <- matrix(c(rep(1:10,10),rep(c(1:5,11:15),each=10)),ncol=2,dimnames=list(paste0('cell',1:100),c('x','y')))
res <- paella(pt,cd)
mat <- matrix(rnorm(1000),nrow=10,dimnames=list(paste0('gene',1:10),paste0('cell',1:100)))
test <- difftest(mat,res)
}
\author{
Wenpin Hou<wh2526@cumc.columbia.edu>, Zhicheng Ji
}
