% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genefit.R
\name{genefit}
\alias{genefit}
\title{Fitting temporal gene expression dynamics}
\usage{
genefit(mat, paella, gene = rownames(mat), interpoint = 200)
}
\arguments{
\item{mat}{A gene expression matrix where each row is a gene and each column is a cell/spot. Row names are gene names and column names are cell/spot names. The column names must be present when fitting \code{paella()}.}

\item{paella}{Direct output from calling \code{paella()}.}

\item{gene}{A character vector of gene names for which pseudotime pattern will be fitted. Must be a subset of row names of \code{mat}.}

\item{interpoint}{A positive integer indicating the number of interpolation points.}
}
\value{
A list containing scaled fitted values, original mean, and original standard deviations.
}
\description{
Fitting temporal gene expression dynamics along pseudotime for each spatial sub-trajectory.
}
\examples{
pt <- setNames(1:100, paste0('cell',1:100))
cd <- matrix(c(rep(1:10,10),rep(c(1:5,11:15),each=10)),ncol=2,dimnames=list(paste0('cell',1:100),c('x','y')))
res <- paella(pt,cd)
mat <- matrix(rnorm(1000),nrow=10,dimnames=list(paste0('gene',1:10),paste0('cell',1:100)))
fit <- genefit(mat,res)
}
\author{
Wenpin Hou<wh2526@cumc.columbia.edu>, Zhicheng Ji
}
