% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescatter.R
\name{genescatter}
\alias{genescatter}
\title{Gene expression scatterplot}
\usage{
genescatter(mat, genefit, paella, gene = rownames(genefit[[1]])[1:6], ncol = 2)
}
\arguments{
\item{mat}{A gene expression matrix where each row is a gene and each column is a cell/spot. Row names are gene names and column names are cell/spot names. The column names must be present when fitting \code{paella()}.}

\item{genefit}{Direct output from calling \code{genefit()}.}

\item{paella}{Direct output from calling \code{paella()}.}

\item{gene}{Character vector of genes to be plotted}

\item{ncol}{A positive integer indicating the number of columns for facets.}
}
\value{
A ggplot2 plot
}
\description{
Plotting the temporal gene expression dynamics across spatial sub-trajectories as scatterplots
}
\examples{
pt <- setNames(1:100, paste0('cell',1:100))
cd <- matrix(c(rep(1:10,10),rep(c(1:5,11:15),each=10)),ncol=2,dimnames=list(paste0('cell',1:100),c('x','y')))
res <- paella(pt,cd)
mat <- matrix(rnorm(1000),nrow=10,dimnames=list(paste0('gene',1:10),paste0('cell',1:100)))
fit <- genefit(mat,res)
genescatter(mat,fit,res)
}
\author{
Wenpin Hou<wh2526@cumc.columbia.edu>, Zhicheng Ji
}
