% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paella.R
\name{paella}
\alias{paella}
\title{Paella: spatial decomposition of cell trajectories}
\usage{
paella(pt, cd, ncores = 1, mode = c("both", "small", "large"), seed = 10)
}
\arguments{
\item{pt}{A named numeric vector of pseudotime. The names indicate different cells or spots.}

\item{cd}{A matrix of spatial coordinates. Each row is a cell or spot. It must have row names indicating the names of cells or spots, and the row names need to agree with the names in \code{pt}.}

\item{ncores}{A positive integer of the number of cores to be used for parallel computing. If 1, no parallel computing will be done.}

\item{mode}{A character vector. The elements must be 'both', 'small', or 'large'. If the length of the vector is larger than 1, Paella will automatically select the optimal mode.}

\item{seed}{A positive integer of random seed}
}
\value{
A list of spatial sub-trajectories assignment for each cell/spot, pseudotime, coordinates, and edges.
}
\description{
Paella takes as input a vector of pseudotime and spatial coordinates of cells or spots. It then identifies multiple spatially distinct while temporally coextending spatial sub-trajectories.
}
\examples{
pt <- setNames(1:100, paste0('cell',1:100))
cd <- matrix(c(rep(1:10,10),rep(c(1:5,11:15),each=10)),ncol=2,dimnames=list(paste0('cell',1:100),c('x','y')))
res <- paella(pt,cd)
}
\author{
Wenpin Hou<wh2526@cumc.columbia.edu>, Zhicheng Ji
}
