% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialplot.R
\name{spatialplot}
\alias{spatialplot}
\title{Paella spatial plotting}
\usage{
spatialplot(paella, pointsize = 1.5)
}
\arguments{
\item{paella}{Direct output from calling \code{paella()}.}

\item{pointsize}{Size of the points shown in the plot.}
}
\value{
A ggplot2 plot
}
\description{
Plotting the spatial sub-trajectories from Paella
}
\examples{
pt <- setNames(1:100, paste0('cell',1:100))
cd <- matrix(c(rep(1:10,10),rep(c(1:5,11:15),each=10)),ncol=2,dimnames=list(paste0('cell',1:100),c('x','y')))
res <- paella(pt,cd)
spatialplot(res)
}
\author{
Wenpin Hou<wh2526@cumc.columbia.edu>, Zhicheng Ji
}
